+++ app/src/main/java/com/example/foosball/Gesture.java
@@ -1,0 +1,39 @@
+package com.example.foosball;
+
+import android.util.Log;
+import android.view.MotionEvent;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.view.MotionEventCompat;
+
+public class Gesture extends AppCompatActivity {
+
+    private static final String DEBUG_TAG = "GAME";
+
+    @Override
+    public boolean onTouchEvent(MotionEvent event) {
+
+        int action = MotionEventCompat.getActionMasked(event);
+
+        switch (action) {
+            case (MotionEvent.ACTION_DOWN):
+                Log.d(DEBUG_TAG, "Action was DOWN");
+                return true;
+            case (MotionEvent.ACTION_MOVE):
+                Log.d(DEBUG_TAG, "Action was MOVE");
+                return true;
+            case (MotionEvent.ACTION_UP):
+                Log.d(DEBUG_TAG, "Action was UP");
+                return true;
+            case (MotionEvent.ACTION_CANCEL):
+                Log.d(DEBUG_TAG, "Action was CANCEL");
+                return true;
+            case (MotionEvent.ACTION_OUTSIDE):
+                Log.d(DEBUG_TAG, "Movement occurred outside bounds " +
+                        "of current screen element");
+                return true;
+            default:
+                return super.onTouchEvent(event);
+        }
+    }
+}